fix error in admin_dledit.php
When editing a download where some files are marked as mirror,
the mirror checkbox will not be checked due to incorrect HTML
syntax for the input element (caused by a typo). As a result
all "mirror" files in the edited download will be saved as
non-mirror files with the next update - unless the mirror flag
is set manually by the person editing the download.

The commit fixes that bug. It also fixes some minor spelling
mistakes in the script and removes the cause of a PHP notice.