remove call to exec() and replace it by alternative functions
The exec() function is usually disabled in most hosted environments,
because of security concerns. Furthermore the call to 'whoami' won't
work on Windows servers running PHP, so we should use a more portable
way of determining the current user's name.

Yes, people should usually avoid Windows as a server system OS, but
nevertheless there are PHP binaries for Windows, too, so we should
try to support that as well.